refactor(native address): move ownerToNativeAddress util to signer PE-6629
