+++ src/types.ts
@@ -30,6 +30,8 @@ import { CurrencyMap } from './common/currency.js';
 import { JWKInterface } from './common/jwk.js';
 
 export type Base64String = string;
+export type NativeAddress = string;
+
 export type PublicArweaveAddress = Base64String;
 export type TransactionId = Base64String;
 export type UserAddress = string | PublicArweaveAddress;
